Faster, more accurate and more robust regex
Enhances the default regex to support more programming languages while
also increasing the ACCURACY and PERFORMANCE by adding bounds checks
and other correctness checks.

All `\s` (whitespace) selectors are now replaced with higher-performing
and more accurate `[ \\t]` (spaces and tabs). JavaScript lacks Perl's
`\h` selector for "horizontal whitespace", so we MUST be more explicit
here to get the same benefits. Why this change? Because `\s` matches new
lines, which means that the entire matching slows down and breaks since
it also checks across line boundaries, which is very bad. So we MUST use
`[ \\t]` for correctness. And programming only uses spaces or tabs, so
we don't need the additional Unicode Whitespace that Perl's `\h` has.

This new regex looks for one of the following:
(When we say "whitespace" below, we only mean spaces or tabs.)

1. `//` (most languages; is unique enough to not need extra bounds...
   is used for "integer division" in some languages, but that's rare
   in actual code, and wouldn't be a false positive anyway since the
   user would ALSO have to put a tag word after their "//" operator)
2. `#` (Python, etc)
3. `<start of line, OR one or more whitespace>--` (Lua etc; the spaces
   increase speed and avoids false positives from syntax
   such as "i--" or "--i".)
4. `<!--` (XML/HTML)
5. `;`
6. `<either "/", OR "start-of-line followed by zero or more
    whitespace"><1 or more "*">` (this matches multiline comments;
    matches "/***** TODO", "* TODO", etc and completely avoids false
    positives. there are no risks thanks to the strict pattern.)
7. `<start of line>` (this is a strict pattern and ensures that we
    are able to match tags within multiline comments, such as inside
    the XML/HTML `<!-- lines... -->` blocks.)

When one of the patterns above match, we check that it's followed by
ALL of the below:

1. `<zero or more whitespace>` (this handles the fact that some
   people like to write `//TODO` and other people write `// TODO`.)
2. Then `<optional "@">` (this handles the `@todo` syntax used in
   JavaDoc, JavaScript, PHP, etc. but note that we still need to
   extend the default tags with lowercase variants if we want to
   actually match them)
3. Then it matches any of the tags.
4. Then `\b` which performs a boundary check to ensure that nothing
   follows after the matched tag. So "// NOTE:" and "// NOTE" will
   match, but not "// NOTEDHEHE".

Lastly, we have the alternative pattern which looks for a markdown
"todo" list item, as seen before `^[ \\t]*- \\[ \\])` (with the
only modification being the improved whitespace pattern here).